requirements.txt
black==21.9b0
click==8.0.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-vlc==3.0.12118
regex==2021.10.8
Rx==3.2.0
tomli==1.2.1
typing-extensions==3.10.0.2
